requirements.txt
functions-framework==3.*
google-auth
pandas
google-api-python-client
tweepy